Split Document Extraction Logic Into Package
While enduring a nightmare of dependency injection, I came to the
conclusion that the document extraction pipeline was in and of itself
a standalone project from it's API counterpart. For that reason, it
deserved to be within it's own repo. However, this introduced the
issue of importing and using it's functionality within the API. To
solve this, I've begun the work to turn this into a standalone python
package, importable using pip and hosted on github.

The following commits will be messy as the debugging process begins,
they will all reference this commit message.